removed news deletion while processing feeds in DefragmentHelper
this is return to previous logic. Feed should store some news as HIDDEN to prevent reaading news on each  feed reload. Later this is an be introduced again but only for deleting news with state DELETED. HIDDEN news will be still in db until they will be explicitly marked as DELETED by some retention policy